requirements.txt
beautifulsoup4==4.8.2
lxml==4.5.0
requests==2.22.0
tqdm==4.42.1
urllib3==1.25.8
pytube3==9.6.4